score embedded into lifebar. score temp abs fix

+++ Renderers/UIElements/ScoreMeter.h
@@ -15,11 +15,11 @@
 @class FontString, Texture2D;
 
 @interface ScoreMeter : NSObject <TMLogicUpdater, TMRenderable, TMMessageSupport> {
-	long			m_nCurrentScore;
+	int				m_nCurrentScore;
 	int				m_nTotalSteps;
 	int				m_nTapNotesHit;
 	int				m_nDifficulty;
-	long			m_nMaxPossiblePoints;
+	int				m_nMaxPossiblePoints;
 	int				m_nRoundTo;
 	int				m_nScoreRemainder;
